refactor: Ticket -> LottoShop으로 변경
