+++ __tests__/LottoShop.test.js
@@ -1,5 +1,5 @@
 import { PRICE } from "../src/domain/Price.js";
-import Ticket from "../src/domain/Ticket.js";
+import LottoShop from "../src/domain/LottoShop.js";
 import { divideByUnit } from "../src/utils/count.js";
 
 test("입력받은 금액에 해당하는 개수를 구한다.", () => {
@@ -18,7 +18,7 @@ test("정해진 개수만큼 로또를 생성한다.", () => {
   const lottoCount = 3;
 
   // when
-  const lottos = Ticket.createLottos(lottoCount);
+  const lottos = LottoShop.createLottos(lottoCount);
 
   // then
   expect(lottos.length).toBe(lottoCount);
