--- src/constants/lotto.js
@@ -1,13 +1,0 @@
-export const LOTTO_NUMBER = Object.freeze({
-  MIN: 1,
-  MAX: 45,
-  LENGTH: 6,
-});
-
-export const LOTTO_STATUS = Object.freeze([
-  { RANK: 1, COUNT: 6, REWORD: 2_000_000_000, IS_BONUS: false },
-  { RANK: 2, COUNT: 5, REWORD: 30_000_000, IS_BONUS: true },
-  { RANK: 3, COUNT: 5, REWORD: 1_500_000, IS_BONUS: false },
-  { RANK: 4, COUNT: 4, REWORD: 50_000, IS_BONUS: false },
-  { RANK: 5, COUNT: 3, REWORD: 5000, IS_BONUS: false },
-]);
