--- src/constants/price.js
@@ -1,5 +1,0 @@
-export const PRICE = Object.freeze({
-  UNIT: 1000,
-  MIN: 1000,
-  MAX: 100000,
-});
