+++ src/domain/LottoShop.js
@@ -1,8 +1,7 @@
-import { LOTTO_NUMBER } from "../constants/lotto.js";
 import { getRandomNumber } from "../utils/random.js";
-import Lotto from "./Lotto.js";
+import Lotto, { LOTTO_NUMBER } from "./Lotto.js";
 
-const Ticket = {
+const LottoShop = {
   checkAndAddLottoNumbers: (store, number) => {
     if (store.has(number)) return;
     store.add(number);
@@ -12,17 +11,17 @@ const Ticket = {
     const randomNumberStore = new Set();
     while (randomNumberStore.size < LOTTO_NUMBER.LENGTH) {
       const number = getRandomNumber(LOTTO_NUMBER.MIN, LOTTO_NUMBER.MAX);
-      Ticket.checkAndAddLottoNumbers(randomNumberStore, number);
+      LottoShop.checkAndAddLottoNumbers(randomNumberStore, number);
     }
     return [...randomNumberStore];
   },
 
   createLottos: (count) => {
     return Array.from({ length: count }, () => {
-      const lottoNumbers = Ticket.issueLottoNumbers();
+      const lottoNumbers = LottoShop.issueLottoNumbers();
       return new Lotto(lottoNumbers);
     });
   },
 };
 
-export default Ticket;
+export default LottoShop;
